requirements.txt
fastapi
sqlalchemy
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlmodel 